Moved the ConfigUtil and OutputIntercept types to different packages.
