+++ data-ingestors/src/main/scala/pipelines/examples/ingestor/RecommenderDataIngress.scala
@@ -12,8 +12,8 @@ import pipelines.examples.data._
 import scala.collection.mutable.ListBuffer
 import scala.util.Random
 import scala.concurrent.duration._
-import pipelines.examples.util.ConfigUtil
-import pipelines.examples.util.ConfigUtil.implicits._
+import pipelines.util.ConfigUtil
+import pipelines.util.ConfigUtil.implicits._
 
 /**
  * Ingress of data for recommendations. In this case, every second we
