+++ data-ingestors/src/test/scala/pipelines/examples/ingestor/WineDataIngressTest.scala
@@ -6,7 +6,7 @@ import akka.actor._
 import akka.stream._
 import pipelines.akkastream.testkit._
 import pipelines.examples.data.WineRecord
-import pipelines.examples.util.test.OutputInterceptor
+import pipelines.util.test.OutputInterceptor
 import com.typesafe.config.ConfigFactory
 
 class WineDataIngressTest extends FunSpec with BeforeAndAfterAll with OutputInterceptor {
