+++ data-ingestors/src/test/scala/pipelines/examples/ingestor/WineRecordsReaderTest.scala
@@ -1,7 +1,7 @@
 package pipelines.examples.ingestor
 
 import org.scalatest.{ FunSpec, BeforeAndAfter }
-import pipelines.examples.util.test.OutputInterceptor
+import pipelines.util.test.OutputInterceptor
 
 class WineRecordsReaderTest extends FunSpec with BeforeAndAfter with OutputInterceptor {
